Fix-Refactor-code(test_users)
